feat: create `src/app/ApolloClientProvider.tsx` component
Create `src/app/ApolloClientProvider.tsx` component to provide
an `ApolloClient` instance to use in all React Client Components
to create and send graphql queries and mutations.

Wrapp the `body` tag of the root layout component with
`ApolloClientProvider` component to the React Client Components
be able to access the `ApolloClient` instance to create and send
the graphql queries and mutations.