feat: create the main component in the home page (`HomePagePostsGrid`)
Create `HomePagePostsGrid` component in `src/components/hompage`
folder, which is the main component in the home page and which uses
the server action `getHomePagePosts` to bring the posts for the user and
displays them in a grid and uses `HomePagePostsGridError` to display
a message with refresh option if an error happened while fetching the
posts.

Add `HomePagePostsGrid` component to `src/app/page.tsx` file which
is the root file of the home page and use `Suspense` component of react
to display a fallback while `HomePagePostsGrid` component is fetching
the posts.