feat: create `PostBookmark` component
Create `PostBookmark` component which is a bookmark icon button
used to bookmark posts.
This component uses `usePostBookmark` hook to handle bookmarking
the post when the user clicks on the bookmark icon button and to display
the state of the bookmark icon button is it active (means that the user
has bookmarked the post) or not.