feat: create the reusable `usePagination` hook
Create `usePagination` hook which is a reusable hook
for fetching the data that comes from the server in pagination
model. This hook uses `graphglQueryAction` server action
to fetch the data (page by page) and append them to data
state, And the hook provides some states and function for
dealing with it, the states are `data`, `isLoading`, `error`,
`paginationModel` and `areThereMoreData`;
And the functions to deal the the hook
are `loadMore`, `addData`, `updateData` and `removeData`.