feat: add & remove a prop in `PostCommentsIcon` component
Add `openPostFullCard` prop to `PostCommentsIcon` component,
which is a function for calling it with `true` as the 1st parameter for
opening the full card dialog of the post and scrolling to comments
section,

Remove `scrollToCommentsSectionState` prop from `PostCommentsIcon`
component because it will not be used anymore.

Remove `dontStopPropagation` option from `PostCardIconButton`
inside `PostCommentsIcon` component because it will make an
unwanted effect after `scrollToCommentsSectionState` prop has
removed.