feat: create `UserBookmarks` component
Create `UserBookmarks` component which is the grid
of the bookmarked posts,

Use hook `useUserBookmarks` to get the bookmarked
posts of the user and display them using `PostsGridContainer`
and `PostCard` components.

Use `PostLoadingCard` component to display loading state
UI while the data of the bookmarked posts are being fetched.

Use span element to trigger that the user has scrolled to the
end of posts grid for fetching more posts if there is more.